\name{HW5Pack-package}
\alias{HW5Pack-package}
\alias{HW5Pack}
\docType{package}
\title{Random Letters}

\description{
Only function Random Letters
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HW5Pack}
\packageIndices{HW5Pack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HW5Pack}

Maintainer: \packageMaintainer{HW5Pack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
